Fix verbose output by forcing max-cpus to 1
Verbose output was lost when max-cpus was > 1 due to multiprocessing and output from subprocesses *not* being directed back to the calling process/its terminal.

Fix this the easy way by decreasing teh number of CPUs used when the user requests verbose output. This is probably ok, becasue verbose output is probably only used for diagnostics.