Fix: removed side-effect of rebuilding grid lines in some "getters"
Rebuilding grid lines will wipe out any grid lines added by
OFracGrid.setNaxGlSpacing. Remove the calls to OFracGrid._remakeGridLineLists in
various getXXX calls.

New behaviour is to raise an exception of the grid was not left in a valid state
by a previous operation (like addGridline)

Go-forward practice should be to leave the grid in a valid state at all times,
but this is a TODO.