Update for efficiency
Any messages generated when fractures are truncated, `OFrac.truncate`, are now transformed into strings just prior to printing them out. Thus, no output-formatted strings are made unless they are actually going to be used.

This is particularly helpful (faster) in batch operations to the DFN, like when resizing (shrinking) the domain when many fractures are culled.