Revised "fracture collapse policy"
Added OFracGrid.collapse_policy instance variable, in addition to the module-level __FX_COLLAPSE_POLICY__ variable.

OFrac objects look first to the policy of their network before the module-level variable.

This was introduced because of module import issues. Some OFrac objects were arriving to calling code as OFrac objects and others were arriving as ofrac.ofracs.OFrac objects --- The module-level variable was not universal for both types of object, so a quick fix was to put this in the concrete OFracGrid object instance.